+++ CHANGELOG.md
@@ -4,10 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- `Catalog::builder()` and `catalog.builder_chained()` shortcuts
+- New `examples` directory to showcase DI patterns and integrations
+
 ## [0.9.1] - 2024-08-15
 ### Fixed
 - `Catalog::builders_for_with_meta()` works corectly for chained catalogs
-
 
 ## [0.9.0] - 2024-07-29
 ### Added
@@ -39,7 +43,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   struct Impl;
 
   // Automatically does `.bind::<dyn Iface, Impl>()`
-  CatalogBuilder::new().add::<Impl>().build();
+  Catalog::builder().add::<Impl>().build();
   ```
 
 ## [0.7.2] - 2023-09-04